+++ models/__init__.py
@@ -1,3 +1,3 @@
-from .functions import Record, AddressBook
+from .classes import Record, AddressBook,Birthday,Notes,Email,Phone
 
-__all__ = ["Record", "AddressBook"]
+__all__ = ["Record", "AddressBook",'Birthday','Email','Phone','Notes']
